Rework docker-compose set up
The Analyst and CSP are intended to be run on different nodes, so rework the
compose set up to accommodate this:

* All network addresses and working directories are now  configurable with a
.env file (default.env included as an example)

* Both services now also use "network_mode: host", allowing them to bind to
their nodes network interfaces. then ensures that analyst IDs reflects
the nodes network address rather than an internal docker address

* The shared docker image can now be built with docker compose. This means that
we only have to ship the compose file to each node, and the nodes will be
responsible for building the images.  This is a good enough work around
until we have a docker registery.